Implement basic which key functionality
Create a new VimExtension named "which-key" and register it within
the plugin.xml file to get the project started.

Integrate the "apache log4j" library for some basic logging.

Extract all key mappings for all modes and save them in the singleton
MappingConfig so we can easily extract nested mappings for a
particular key sequence.

Replace the typing handler of IdeaVim with a custom one to intercept
all key strokes and check for nested mappings. If such exist show a
popup Balloon to display all available mappings.