Format nested mapping entries
Strip the previously typed characters to only show the next character
which needs to be pressed to either execute a command or open another
prefix.

Replace the hard to read <space> character with the string "SPC". This
should make corresponding mappings easier to read and identify.

Escape angle brackets (< and >) so they are not interpreted as HTML tags.