Split popup content into columns
For a better readability and to avoid the height of the Balloon
becoming to big we split the mappings entries into multiple columns.

Check for the longest entry and use this as a reference to determine
how many columns can be properly displayed. Then split the mappings
between those columns.

The width calculation is not perfect yet, but it already makes the
plugin usage way more enjoyable.