Implement display delay for the WhichKey popup
To prevent the popup from flickering on fast consecutive key presses
implement a short delay (currently 200ms). The popup will only be
shown if no other key was pressed for the duration of the delay.

To implement this change using Kotlin Coroutines the appropriate
library was added as a dependency as well as the JDK 8 Kotlin standard
library to make it work correctly.