Set minimum Intellij version to 2020.1 (201)
Set the "version" parameter for Intellij to 2020.1 to build the plugin
with this specific version of Intellij.

Use the option "updateSinceUntilBuild" to avoid that the Intellij
version gets automatically patched into the plugin.xml file as since
and until build number.

Set the since-build number within the plugin.xml file to 201, as this
is the minimum Intellij version which is needed by IdeaVim version 0.62.
Do not specify an until-build number to not restrict the plugin for
higher versions of Intellij.