Enable popup for non-character keys strokes
In order to enable the plugin for non-character key strokes (e.g.
escape, tab, ctrl + <char>, alt + <char>, etc.) we overwrite the IdeaVim
'VimShortcutKeyAction' action with a custom class. The custom class
handles the popup display before passing everything to the default
class of IdeaVim.

Intellij will complain about the registration of an action id which
was already registered before but besides this there are no
consequences in doing so.

To avoid duplicate code the new PopupConfig object now handles the
display of the popup in one central place.