Make command and prefix descriptions configurable
Make the description for every command or prefix configurable in the
.ideavimrc file by defining a corresponding variable. To identify all
variables related to the configuration of any descriptions they need
to start with a common prefix:

g:WhichKeyDesc_

The name is only for the variable filtering, while the value holds the
key sequence as well as the description. Since IdeaVim does not
support any collection variables, there has to be one variable per
combination of a mapping and a description.

Also add some basic information (and images) into the README file:
- example GIF
- manual installation
- customization of descriptions