Fix column spacing
Reduce the frame width used for calculations to 65 percent.
This is done because the HTML content in the balloon takes up more
space than anticipated (probably due to spacing and padding).
I could not find the exact code for it so the factor 0.65 is a guess
based on experimenting.

Also always use the full available frame width as display area. This
is archived calculating a column width instead of relying on the
maximum string width. Also add a missing 'px' to the width attribute.