Fix prefix comparison for the '<' character
Compare the prefixes of nested mappings not by strings but instead use
a sub list of the key-strokes instead. This makes the comparison more
accurate while not affecting the performance in any measurable way.

Without the fix typing the '<' character will display prefixes like:

- Space>
- Alt-n>
- C-n>

This can happen to all special or modified keys which are enclosed in
angle brackets.