Dismantle key and description caching
Reload nested mappings and descriptions on every key press instead of
cache it all on plugin startup. This way every change and reload of
the .ideavimrc is immediate reflected within Intellij. This makes
testing and tweaking the settings way faster as it does not need a
restart for every change.

Obviously calculating all nested mappings newly on every key press is
less performant than the cached solution. But from a direct comparison
the difference are not that significant.