Display default VIM actions (gg, zz, etc.)
If enabled the plugin will display default VIM actions like 'gg',
'zz', etc. within the popup. For this we extract all such mappings at
startup and if enabled add them to the ones from the .ideavimrc file.

The most recent IdeaVIM update (0.65) just came out and changed the
'Node' classes which we check to extract the default actions. Now
these are parameterized and we have to pass the used type to them.
Therefore we have to update the needed IdeaVIM version to match this
new requirement.