Don't display popup for VIM motions
The follow up argument for certain VIM motions/actions could trigger
the popup if there were any mappings registered for the argument key.
Displaying a popup here is wrong as the key press was already handled
and therefore subsequent key presses would not lead to the shown
actions being executed.

For example "fg" should not display any mappings for "g" but instead
just jump to the next "g" character.

To fix this issue we check if the CommandBuilder is waiting for a
DIGRAPH argument and if so skip checking for mappings and displaying
the popup completely.