Refactor maximum width calculation
Refactor the FormatConfig object to use JLabel HTML formatting instead
of separate JLabels. This will ensure that the calculated width is
exactly the same as in the popup instead of being just an
approximation.

By refactor this we can also remove a lot of special "calculation
functions" that we had implemented before. This way we have no width
related code in the FormatConfig object.