Consider the time needed for calculations in the popup delay
Subtract the time needed for calculations of the nested mappings and
available width etc. from the popup delay. We do this to provide a
consistent delay time independent from the time needed for everything
else.

Especially since the first calculations tend to need more time than
the following ones this makes the overall experience more consistent.