Allow setting descriptions for built-in vim commands
without having to show each and every possible vim command there is.

The behaviour when `g:WhichKey_ShowVimActions` is `true` is unchanged,
but if it is not `true``, the new behaviour includes all built-in commands
that the user set a description for (otherwise the command is not shown
because the description is empty).

This allows for users to selectively enable showing hints for certain
built-in vim commands without cluttering the which-key pop-up.